Player Camera Follow + WIP Ship Station System
WIP Ship Station: Added public vars isAttached and attachedStation, change on station collision.
