--- CS377-DROOO/Assets/Grounded.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 81077917715366648834db3021133308
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
