+++ CS377-DROOO/Assets/Scripts/playerCamera.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class playerCamera : MonoBehaviour
+{
+    public Transform player;
+
+    void FixedUpdate()
+    {
+        transform.position = new Vector3(player.position.x, player.position.y, transform.position.z);
+    }
+}
