+++ CS377-DROOO/Assets/Scripts/playerCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1c934438e3489fb4ea78149f96e3378f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
