+++ CS377-DROOO/Assets/Scripts/stationControl.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b477e0f481b87443b924b16053f9427
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
