--- CS377-DROOO/Assets/playerMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a2ff4c526bf5d6a4a8c9ef9902239e5f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
