+++ CS377-DROOO/CS377-DROOO.sln
@@ -1,0 +1,20 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 11.00
+# Visual Studio 2010
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assembly-CSharp", "Assembly-CSharp.csproj", "{6A4B5E11-2B4B-4B6D-BAAE-1365F7A9AD51}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6A4B5E11-2B4B-4B6D-BAAE-1365F7A9AD51}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6A4B5E11-2B4B-4B6D-BAAE-1365F7A9AD51}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6A4B5E11-2B4B-4B6D-BAAE-1365F7A9AD51}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6A4B5E11-2B4B-4B6D-BAAE-1365F7A9AD51}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
