+++ CS377-DROOO/Library/StateCache/SceneView/2bccca-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":-1.4257965087890626,"y":1.0609962940216065,"z":30.21973419189453},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":16.91865348815918,"orthographic":true}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":-3.666837692260742,"y":2.438009023666382,"z":-19.226810455322267},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":13.676725387573243,"orthographic":true}
